# spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "LunchGroupAPI -kyewonseo",
    "description": "in Solidware",
    "contact": {
      "email": "kyewon76@gmail.com"
    }
  },
  "host": "localhost:3333",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/v1/groups": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "사람들을 무작위로 그룹으로 나눈다",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "minSizeGroup",
            "in": "query",
            "description": "default minSizeGroup=2",
            "required": false,
            "type": "number"
          },
          {
            "name": "numberOfGroup",
            "in": "query",
            "description": "default numberOfGroup=3",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GroupRes"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorRes"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Get all group",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GroupRes"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorRes"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ErrorRes": {
      "type": "object",
      "properties": {
        "responseStatus": {
          "type": "integer",
          "description": "실패 코드"
        },
        "responseMsg": {
          "type": "string",
          "description": "실패 이유"
        }
      }
    },
...
    "GroupRes": {
      "type": "object",
      "properties": {
        "responseStatus": {
          "type": "integer",
          "description": "성공여부"
        },
        "responseMsg": {
          "type": "string",
          "description": "응답 메시지"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "description": "생성된 group id"
              },
              "name": {
                "type": "string",
                "description": "그룹이름"
              },
              "persons": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "description": "생성된 person id"
                    },
                    "name": {
                      "type": "string",
                      "description": "이름"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}